+++ CHANGELOG.md
@@ -74,6 +74,7 @@ Methods still under consideration in `shiny.experimental.ui`:
 * Fixed #676: The `render.data_frame` selection feature was underdocumented and buggy (sometimes returning `None` as a row identifier if the pandas data frame's index had gaps in it). With this release, the selection is consistently a tuple of the 0-based row numbers of the selected rows--or `None` if no rows are selected. (#677)
 * Added tests to verify that ui input methods, ui labels, ui update (value) methods, and ui output methods work within modules (#696).
 * Adjusted the `@render.plot` input type to be `object` to allow for any object (if any) to be returned (#712).
+* In `layout_column_wrap()`, when `width` is a CSS unit -- e.g. `width = "400px"` or `width = "25%"` -- and `fixed_width = FALSE`, `layout_column_wrap()` will ensure that the columns are at least `width` wide, unless the parent container is narrower than `width`. (#772)
 
 ### Other changes